refactor: promote the provider-proxy-set cluster to its own component
Five siblings shared the `provider-proxy-set-` prefix and were facets of one
bounded component, which is the subdivision trigger stated in
.claude/rules/design-philosophy.md. This branch's work made it acute: the
lifecycle passed thirteen hundred lines and had also become the home of the
decision vocabulary it prints and of the formatter for an identity it does not
own, so following a terminalization import meant opening two files to understand
one printed reference. code-critic scored the result below the elegance gate.

The cluster is now `services/provider-proxy-set/` with the prefix stripped: the
live transition state machine is `index.ts` as the rule prescribes, the reference
formatter sits beside the identity it formats, and the decision vocabulary and
rendering have their own facet. The state machine stays whole — splitting private
transitions to shorten a file would trade cohesion for line count.

Behaviour is unchanged; this is a move and a re-home. The recovery-policy
invariant's boundary authorizations were updated to the new paths, with every
justification and entry intact — no allowlist gained a member and no assertion
was weakened. Its fingerprint ordering shifted only because `index.ts` sorts
before `inheritance.ts`.
