+++ docs/core-modules.md
@@ -103,7 +103,7 @@ Projection freshness is not authority. A Corpus commit remains durable even if a
 
 The coordinator layer owns process lifecycle, the three-era boot sequence (Kernel → Recovery → Runtime Health Components), projection-consumer wiring, provider-host coordination, job-backed KB source import/reindex proxying, and the corpus notify seam. Its provider-host administration service is an aggregator and router, not the sole host authority: the coordinator-local manager and every live provider proxy report their own live/tombstoned inventory and execute their own admission and eviction. Federation is all-or-error, and exact owner revalidation prevents a selection race from retargeting a destructive operation. It also owns the evidence-gathering half of carrier observation that the classifier deliberately does not: `composition/carrier-observation.ts` gathers local evidence (admission state, the OS process table, and the durable-CLI identity row), classifies health and idle synchronously from that evidence, and builds wait's asynchronous observation port. For inherited wait-time operations only, `live/carrier-observer.ts` adds a bounded `operation.status.v1` pass. That method is a nonce- and proxy/build-bound, read-only ledger-liveness handshake; `operation.inspect.v1` remains the richer durable-saga ambiguity surface used by reconciliation. An unanswered status question is `unknown`, not absence, and every consumer treats it as conservatively active. Only a positive `absent` may render a wait interruption, and even that derived finding authorizes no recovery, cleanup, reconciliation, terminal materialization, or other mutation. `tests/invariants/no-carrier-observation-in-action-paths.test.ts` enforces that boundary at every import depth. Recorded process containment is narrower and independent: `infra/process-containment.ts` revalidates the recorded group leader and process start time before signalling, then requires stable absence before reporting success. Its success means no verified recorded target remains; it cannot prove original group members disappeared after leader-PID reuse. Proxy callers can include recorded provider roots; coordinator-local host teardown supplies an empty root set and reaps only its recorded group. The coordinator layer is the only place allowed to compose multiple parent-daemon domains together and the only place that speaks to both transport and domain owner modules/contracts at once. `coordinator/handoff-runner.ts` owns the cross-build handoff decision and execution — probing a live incumbent's health, resolving the routing decision, and spawning/delegating into a validated newer build — shared by CLI pre-flight delegation, `wait jobs` cursor resume, and backend cold-start handoff (`src/coordinator/bootstrap.ts` calls it when startup finds a newer active-store selection).
 
-Authorization of the explicit provider-proxy containment path is coordinator-service policy. `coordinator/services/provider-proxy-authority-fault.ts` owns the control-call policy types plus the separate non-consuming incident and terminal-fault channels; `provider-proxy-operation-activation.ts` requires every mutation to declare `indeterminate: 'retry-safe' | 'requires-containment'` and classifies each failed control call; `provider-proxy-set-lifecycle.ts` owns the preserve/drain/stop-and-reap decision, records that decision before action, and carries it to the reap boundary. A claim-bearing reap requires a containment-qualified terminal fault, while faultless retirement can reap only with literal `liveClaims: 0`. The provider-proxy domain owns enforcement and reap mechanics for that authorized decision; the separate deadline machines in guardian and reaper retain autonomous enforcement after control loss.
+Authorization of the explicit provider-proxy containment path is coordinator-service policy. `coordinator/services/provider-proxy-authority-fault.ts` owns the control-call policy types plus the separate non-consuming incident and terminal-fault channels; `provider-proxy-operation-activation.ts` requires every mutation to declare `indeterminate: 'retry-safe' | 'requires-containment'` and classifies each failed control call; `provider-proxy-set/decisions.ts` owns the preserve/drain/stop-and-reap vocabulary and pure log rendering, while `provider-proxy-set/index.ts` records each decision before action and carries it through the live transition state machine to the reap boundary. A claim-bearing reap requires a containment-qualified terminal fault, while faultless retirement can reap only with literal `liveClaims: 0`. The provider-proxy domain owns enforcement and reap mechanics for that authorized decision; the separate deadline machines in guardian and reaper retain autonomous enforcement after control loss.
 
 ## Runtime Components
