+++ tests/integration/provider-proxy/operation-lifecycle.integration.test.ts
@@ -60,7 +60,7 @@ import {
   MAX_PROVIDER_REPLAY_BYTES,
   MAX_PROVIDER_REPLAY_EVENTS,
 } from '#src/provider-proxy/ledger.js';
-import { proxyHandoffRedeemResultSchema } from '#src/coordinator/services/provider-proxy-set-inheritance.js';
+import { proxyHandoffRedeemResultSchema } from '#src/coordinator/services/provider-proxy-set/inheritance.js';
 import { PROXY_CONTROL_HEARTBEAT_MS, PROXY_CONTROL_LEASE_MS } from '#src/provider-proxy/orphan-deadline.js';
 import {
   decodeProxyControlFrame,
